requirements.txt
aiofiles==23.2.1
alembic==1.13.0
aniso8601==9.0.1
annotated-types==0.6.0
anyio==3.7.1
async-exit-stack==1.0.1
async-generator==1.10
autopep8==2.0.4
bcrypt==4.1.1
Flask==3.0.0
pandas==2.0.3
influxdb-client==1.36.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.105.0
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.2
h11==0.14.0
httptools==0.6.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
maturin==1.4.0
orjson==3.9.10
passlib==1.7.4
promise==2.3
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
Rx==3.2.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
toml==0.10.2
typing_extensions==4.9.0
ujson==5.9.0
urllib3==2.1.0
#uvicorn==0.24.0.post1
watchgod==0.8.2
websockets==12.0
pickleshare==0.7.5
numpy==1.23.5
uvicorn== 0.24.0.post1
